Completed Create Message Test

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -1,10 +1,12 @@
 package com.example.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entity.Account;
 
-public interface AccountRepository extends JpaRepository<Account, Long>{
+public interface AccountRepository extends JpaRepository<Account, Integer>{
 
     Account findByUsername(String username);
 }
